% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{communities}
\alias{communities}
\title{Canadian Communities from OpenStreetMap}
\format{
\subsection{\code{communities}}{

A tibble data frame with 11417 rows and 6 columns:
\describe{
\item{name}{Name of community. Character.}
\item{type}{Type of community. Either "city", "town", "village", or "hamlet". Factor.}
\item{prov_terr}{Abbreviation of province or territory. Factor.}
\item{fcst_zone}{English name of forecast zone. 47 communities are not in any forecast zone, so they are assigned \code{NA}. Factor.}
\item{lng}{Longitude of the community. CRS: WGS84}
\item{lat}{Latitude of the community. CRS: WGS84}
...
}
}
}
\source{
\url{https://openstreetmap.org}
}
\usage{
communities
}
\description{
All cities, towns, villages and hamlets in Canada from OpenStreetMap.
}
\keyword{datasets}
