% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indigenous_lands}
\alias{indigenous_lands}
\title{Aboriginal Lands of Canada Legislative Boundaries}
\format{
\subsection{\code{indigenous_lands}}{

An \code{sf} tibble data frame with 13 rows and 5 columns:
\describe{
\item{type}{Boundary type. Either "Indian Land", "Indian Reserve", or "Land Claim".  Factor.}
\item{name_en}{English name of area. Character.}
\item{name_fr}{French name of area. Character.}
\item{prov_terrs}{Abbreviation of province/territory the area is located in, or multiple (seperated by commas) if it covers multiple provinces/territories.}
\item{fcst_zones}{English name of the forecast zone the area is located in, or multiple (seperated by commas) if it covers multiple. 8 polygons are not in any forecast zone, so they are assigned \code{NA}. Character.}
\item{geometry}{Indigenous lands legislative boundary polygons. \code{sf} sfc_GEOMETRY.}
...
}
}
}
\source{
\url{https://open.canada.ca/data/en/dataset/522b07b9-78e2-4819-b736-ad9208eb1067}
}
\usage{
indigenous_lands
}
\description{
Locations and names of all Aboriginal land legislative boundaries in Canada from OpenData Canada.
}
\details{
Polygons were simplified using \code{rmapshaper::ms_simplify()} to reduce file size,
resulting in less details for some borders.
}
\keyword{datasets}
