% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_presence_in_polygon.R
\name{mark_presence_in_polygon}
\alias{mark_presence_in_polygon}
\title{Mark presence of points/polygons in polygons}
\usage{
mark_presence_in_polygon(x, y, id_col = "id")
}
\arguments{
\item{x}{A sf object containing the points/polygons to be tested.}

\item{y}{A sf object containing the polygons to be tested against.}

\item{id_col}{The column name in \code{y} that contains the ids of the polygons.}
}
\value{
A data frame with the same columns as \code{x}, and an additional column containing the id(s) of the polygon(s) in \code{y} that each \code{x} is within.
}
\description{
Adds a column to \code{x} labelling the \code{id_col} of \code{y} that each \code{x} is fully within,
or the combined \code{id_col} seperated by a comma if it is partially within multiple polygons.
}
