% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecast_zones}
\alias{forecast_zones}
\title{Canadian Public Forecast Zones}
\format{
\subsection{\code{forecast_zones}}{

An \code{sf} tibble data frame with 13 rows and 5 columns:
\describe{
\item{prov_terr}{Abbreviation of province(s)/territory(ies) the zone covers, seperated by commas where more than 1. Character.}
\item{name_en}{English name of the zone. Factor.}
\item{name_fr}{French name of the zone. Factor.}
\item{perimeter}{Perimeter of the zone. Units: km.}
\item{area}{Area covered by the zone. Units: km^2.}
\item{geometry}{Forecast zone boundary polygons. \code{sf} sfc_GEOMETRY.}
...
}
}
}
\source{
\url{https://api.weather.gc.ca/collections/public-standard-forecast-zones}
}
\usage{
forecast_zones
}
\description{
The Public Standard Forecast Zones layer is a collection of public program
forecast location zone polygons that represents bounded measurable locations
at the Public program Standard level.
The public program standard level is used in most
forecasts, warnings, watches, advisories and special weather statements.
}
\details{
Polygons were simplified using \code{rmapshaper::ms_simplify(keep_shapes = TRUE)} to reduce file size,
resulting in less details for some borders.
}
\keyword{datasets}
