% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces_and_territories}
\alias{provinces_and_territories}
\title{Canadian provinces from OpenStreetMap}
\format{
\subsection{\code{provinces_and_territories}}{

An \code{sf} data frame with 13 rows and 5 columns:
\describe{
\item{abbreviation}{Abbreviation of province or territory. Factor.}
\item{name_en}{English name of province or territory. Factor.}
\item{name_fr}{French name of province or territory. Factor.}
\item{is_province}{TRUE if province, FALSE if territory. Logical.}
\item{geometry}{Province/territory boundary polygons. \code{sf} sfc_GEOMETRY.}
...
}
}
}
\source{
\url{https://openstreetmap.org}
}
\usage{
provinces_and_territories
}
\description{
Locations, names, and abbreviations of all provinces and territories in Canada from OpenStreetMap.
}
\details{
Polygons were simplified using \code{rmapshaper::ms_simplify()} to reduce file size,
resulting in less details for some borders.
}
\keyword{datasets}
